modeling/requirements.txt
torch>=2.0
torchvision
pytorch-lightning
opencv-python
pillow
numpy